Presents more helpful help.
We've gone with optimist rather than docopt because docopt automatically handles lots of stuff. Whisper interface is rather simple and free-form, so it doesn't quite make sense. With optimist being more of a 'raw parser', we can easily handle the options the way we want.

bonus points: tasks now get the options we've gathered too.