Breaking change 0.3: Support package.json configs.
As these configurations now lie in the `whisper` object in the
environment, the command line arguments for a task have been moved
exclusively to the `command` field.

Configurations and plugins are loaded either from the user .whisper
configuration or the local project configuration.

Fixes #6, and it's a step in the direction of #11

--HG--
extra : amend_source : 61baac9bab7f67ad859d8aaf119b8b6e1b0992ce